// apps/next-client/src/lib/store/snackbar/types.ts
import type { SnackbarProps } from '@/lib/components/Snackbar/Snackbar'

export interface SnackbarState {
    queue: SnackbarProps[]
    current?: SnackbarProps
}

export enum SnackbarActionTypes {
    CLEAR_QUEUE = 'snackbar/clear',
    GET_NEXT = 'snackbar/next',
    PUSH = 'snackbar/push',
    CHECK_PARAMS = 'snackbar/check_params',
}